LPS-113561 Avoid declaring the `exportRecordSet` function as a global
As part of [LPS-113561](https://issues.liferay.com/browse/LPS-113561) we
want to eventually deprecate the `Liferay.provide` function.

In this change, we avoid declaring the `<portlet:namespace
/>exportRecordSet` function on the `window` object and remove the
unnecessary AUI dependencies.

To test this change:

- Navigate to "Content and Data > Dynamic Data Lists"
- Click on the "+" button to create a new "Dynamic Data List"
- Enter a name (for example Dynamic Data List 01)
- Click on the "Select" button underneath the "Data Definition" input field -
- Click on of the items in the list (for example "Contacts")
- Click on the "Save" button
- Click on the "kebab" menu on the right of your Dynamic Data List
- Choose the "Export" item
- Click on the "Save" button in the modal window

As a result, a .csv file should be downloaded by your browser and no
errors should appear in the browser's console.