LPS-113561 Avoid unnecessary use of `Liferay.provide` in dynamic-data-lists-web
As commented [here](https://github.com/wincent/liferay-portal/pull/320#discussion_r436726226),
the use of `Liferay.provide` is not necessary in this case and can be removed.
Additionally, this function doesn't need to be prefixed with the `<portlet:namespace />`.

To test this change:

- Navigate to "Content and Data > Dynamic Data Lists"
- Click on the "+" button to create a new "Dynamic Data List"
- Enter a name (for example Dynamic Data List 01)
- Click on the "Select" button underneath the "Data Definition" input field -
- Click on of the items in the list (for example "Contacts")
- Click on the "Save" button
- Click on the "kebab" menu on the right of your Dynamic Data List
- Choose the "Export" item
- Click on the "Save" button in the modal window

As a result, a .csv file should be downloaded by your browser and no
errors should appear in the browser's console.