LPS-113561 Avoid Liferay.provide in document-library-web
As part of [LPS-113561](https://issues.liferay.com/browse/LPS-113561) we
eventually want to deprecate the `Liferay.provide` function

To test the change for the `<portlet:namspace />selectFileEntryType`
function:

- Navigate to "Content and Data > Documents and Media"
- Click on the "+" button and choose "Folder"
- Type a name for the folder you want to create
- Click on the "Save" button
- Click on the kebab menu to the right of the folder you created and
  choose "Edit"
- Select the radio button with the label
  "Define Specific Document Type Restrictions and Workflow for This
   Folder (A folder)"
- Click on "Select Document Type"
- In the modal window choose one of the document types and click on
  "Choose"
- Click on the "Save button"

As a result, a notification should be displayed letting you know your
request has completed successfully and no JS errors should appear in the
browser's console.

To test the change for the `<portlet:namspace />updateMultipleFiles`
function:

- Navigate to "Content and Data > Documents and Media"
- Click on the "+" button and choose "Multiple Files Upload"
- Click on the "Select Files" button
- Upload at least 2 files (images or text documents are OK)
- Click on the "Publish" button

As a result, the files you uploaded should be visible in the "Documents"
list and no JS errors should appear in the console.