LPS-115204 Prevents null pointer when looking for the triggerLabel
This can happen under 2 conditions:
- There's no active item inside the `navigationItems` array
- The `navigationItems` array is empty (thus forcing condition 1)

This simply bypasses the JS condition. In small screens, the trigger
will show (empty) and won't expand a dropdown since it has no elements.

First thought that comes to mind is to add checks inside the component
to prevent rendering if no items are passed, which sounds reasonable but
would require some additional thought and testing on the matter.