LPS-115484 Make path to CSS variables file configurable
First, we will use the Tokens-Path bundle header to get the path to the
JSON file inside the WAR and will use WEB-INF/tokens.json as default
value.

Then we will change the format of the JSON file to make it more
extensible and maintainable in the future.

Note that themes are WAR files, not OSGi bundles, so people need to add
the Tokens-Path header to the liferay-plugin-package.properties file to
make it appear in the MANIFEST.MF file of the generated OSGi bundle.

Also note that we check for the existence of a </theme> tag in the
liferay-look-and-feel.xml file to distinguish theme WAR files from
other type of WAR artifacts.