LPS-119754 Assert keys for operational system compability and prevents the key event default behaviour just when pressing the combination of the keys to open the menu
In this commit I moved the `event.preventDefault` from the begging of the method to when detecting if the keys match to the hotkey. It will make the user able to continue reloading pages using  "CMD + R"  on the Browser. Also, there is a different key for `m` character on Windows OS, it will uppercase the letter since the shift its pressed and It was fixed just normalizing the event.key return