LPS-127081 Move useLiferayState into frontend-js-react-web
This requires us to set up TS compilation in frontend-js-react-web.

Obviously, all of this is horridly manual at this point and just serves
as proof that you _can_ require TS from another module and have it work
ie.:

- Build process in either module picks up type errors.
- Test runner in either module runs both JS and TS.
- LSP-powered editor features (eg. display of type info, errors) works.

I'm going to upload this just as a proof of concept and continue to
investigate more ergonomic ways of configuring this. The main pain point
is that ambient types don't get picked up automatically, which is why I
added that "triple-slash directive" to force their inclusion:

    https://www.typescriptlang.org/docs/handbook/triple-slash-directives.html

Other things tried that didn't work included adding the other project
under the `typeRoots` config, and renaming the type folder to match the
pattern that TS seems to expect for named scope packages (eg.
`liferay__frontend-js-state-web`); I also tried just using the named
scope literally, but that didn't work either.

Next thing I'll be investigating is leveraging project references in
some way:

    https://www.typescriptlang.org/docs/handbook/project-references.html