LPS-127081 Port frontend-js-react-web index to TS
This shows how to do a partial port from JS to TS; in this case, all of
the imported modules continue to be vanilla JS, and `useLiferayState` is
TS. By converting the file to TS, external projects will be able to
"see" the types and make use of them.

Added `"composite": true` in order to be able to test this with `yarn
run tsc --emitDeclarationOnly`, and also because it is a prerequisite to
actually being able to use this from any other module.