LPS-127081 [DEMO-ONLY] Show cross-module TS working
Just a couple of components that share state with each other via
the `Liferay.State` (indirectly, via the `useLiferayState()` hook).

We don't need to actually merge this, seeing as the clay-sample-web
widget should probably stay exactly as it is. Just showing that with all
the config changes in this PR, stuff still works.