LPS-127081 Set up "main" field for frontend-js-state-web
Because we need it if we want our magical auto-setup of Jest
`moduleNameMapper` to pick up this module.

This will allow test in frontend-js-react-web to pass when we do:

    import {State} from '@liferay/frontend-js-state-web';

As opposed to spelling out the path fully:

    import State from '@liferay/frontend-js-state-web/src/main/resources/META-INF/resources/State';

which is what we were doing in an earlier iteration of this PR:

    https://github.com/wincent/liferay-portal/pull/375

I will still need to sent a PR to `npm-scripts` to close the loop here,
because it currently assumes `.js` (well, technically it assumes
whatever is in the `"main"` field of the package.json, but that has to
be ".js" in order for everything else to work; our
bundler/loader/registry assume that "main" is always an ES5-ish file
ready to be served up), but in the meantime I hackily edited my
`node_modules` folder and saw this working.