LPS-127084 [POC] Model `destroyPortlet` events in terms of global state
Just putting this here for discussion. Context:

- https://issues.liferay.com/browse/LPS-127084
- https://liferay.slack.com/archives/C3JBR21HA/p1617959223380700

Some concerns are expressed inline in comments in here and in JIRA. The
comment that I allude to about SPA on LPS-127084 is actually this one:

- https://issues.liferay.com/browse/LPS-127083?focusedCommentId=2382526&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-2382526

Which we then went on to discuss in some more detail in this Google doc:

- https://docs.google.com/document/d/1T1KqhUNar5ze7JI-xhWX0fLZ3IAZxeS0XRRYy1--YtA/edit

So, with all that out the way; this is just to show one possible/minimal
shape that a global-state-based model for portlet lifecycle could begin.
Doing this piecemeal (on event at a time) may not be the right approach
as it doesn't feel like a step forward (more like a lateral step). It
may be that we should consider a more dramatic overhaul, but that is
also pretty difficult, if we're anchoring it in the context of a legacy
AUI module that is loaded by a pretty esoteric mechanism, _and_ happens
to be mostly consumed from JSPs, which have their own legacy baggage.

Anyway, not for merging; just for discussion.