LPS-130155 Support legacy configuration values
Note that we need to change ThemeDisplay to HttpServletRequest in the API because
the legacy values could be stored in the session (as PortalPreferences).

This is a bit unfortunate, because ThemeDisplay is more specific than
HttpServletRequest, but since both are very generic in any case, seems like
a valid trade-off. In any case, we may add more specific factory methods in the
future if we want make the API request agnostic.