DO NOT MERGE !!!!! DO NOT MERGE !!!!! DO NOT MERGE !!!!!
This is an example to demonstrate how the taglib would be used. It uses example.css,
which lives inside frontend-taglib (I have used this bundle because it was handy,
but it could be any other), from Document Library portlet.

To see how it works just add a Document Library portlet to the home page and see how
the text renders in red.