LPS-136999 Show portlet name in dataset
This is so that users can pass custom element attributes with values based on
the portlet name.

For example, one could configure a custom element attribute containing the
portlet name so that the portlet can use it to implement a client side routing
strategy based on the portlet's automatic friendly URL mapping.

So, for instance, if someone registers a custom element portlet based on
<vanilla-counter> custom tag, she can then define an attribute like this:

router-base=vanilla_counter_40149

To then navigate to URLs like:

http://localhost:8080/web/guest/home/-/vanilla_counter_40149/s/normal

From the JavaScript code.