LPS-X Adds an initial global Liferay interface
This is going to be arguably the hardest thing to get right in our migration
to TypeScript. The `Liferay` object is global and augmented from way too
many different sources making it really hard to type.

This commit aims to start simple and get us to a point where we can start
adding more types as we go. For now, it simply adds the `Liferay` object
and describes all the properties defined in `portal-web/top_js.jspf`.

It is left for future commits to:
- Refine existing types to better match the `Liferay` object
- Add more types as needed
- Figure out a better way to handle the `Liferay` object