LPS-X Converts some trivial modules to TS
I picked the lowest hanging modules I could find to get them out of the
way.

One interesting thing to note is that we have some argument type guards
in place. These usually take the form of `if (typeof arg !== <type>)`.
We usually check and throw an Error during runtime if the types don't
match.

This should no longer be necessary if we could guarantee all of our consumers
were using TS. Unfortunately, that's not really the case, so the guards
can still play a role albeit with diminishing returns.

A side effect of this is that when we're testing the guard, TS will complain
automatically about us passing the wrong type. That is, if we try to pass
an object to a method that expects a number, TS will complain about it. The
solution picked here is to use `as any` to cast the argument to get it through.