LPS-X Converts utility navigate to TS
The most interesting part in this conversion is the code change to use
`instanceof` instead of a custom if clause to guard the type and narrow
down the url argument to the correct type.

This is a rather common pattern in Liferay. Accepting several different
arguments and then using the same variable indistinctly after some
normalization steps.

A possible alternative would be to create a real "normalizer" function
`(url: string | URL) => string` that would take the url and return a
normalized version with the type narrowed down to `string`.