LPS-X Converts another batch of semi-trivial conversions to TS
Interesting note here though is the fact that TS caught a bug in our test
or potentially in our implementation of `addParams`.

The implementation checks wether the first argument is `string | object`. The
test goes on to put that to the test passing an array which, coincidentally
is an object as far as `typeof` is concerned. The test expects a TypeError
to be thrown which funny enough, is thrown because the call is missing the
second argument.

I've decided to remove the test assertion since passing an array as params
sounds like a valid use case to me even though the keys would be the number
indexes.