LPS-X Updates form utilities to TS
- [HTMLCollection should be string indexable](https://github.com/microsoft/TypeScript/issues/19437).
But it isn't, so `form.elements[elementName: string]` is a no go. I decided
to go with `HTMLFormControlsCollection.namedItem()` instead which should
do about the same.
- Improves `isObject` help the compiler infer the type resulting after it
by returning a non-nullable version of same type passed to it. However, it
won't return `never` for `isObject(foo: string | boolean | number)` so it
can still use some further refinement.