LPS-139645 Bundles remote-app-client-js module but prevents it from rendering unnecessarily
This is meant for client apps while the part that takes care of the
communication server-side is the remote_protocol/bridge module that
lives in the remote-app-web module and gets instantiated whenever an
iframe-type of remote app is used.

Bundling this by default means that apps could fetch the client sdk direcly
from their Liferay installation. While this is not the most performant
way, it provides a simple way to access it and keep versions stable regardless
of future changes in the sdk (although we intend to keep APIs in there
backwards compatible forever).