LPS-141185 Adds the possibility to pause the execution of the scheduler
It adds the possibility of pausing and continuing the execution of the scheduler, this is important for the singleton strategy, that during the component unmount we pause the execution of the queue until the next component is mounted to continue execution. That's because we reuse the same instance when browsing with SPA and we can't prevent the component from being mounted and unmount.