LPS-141185 Render A11y on `top_head` instead of bottom
Going back again to render the tool on the top head, before we had switched to the bottom because we had race condition problems with the SPA implementation that wrapped the component inside a surface causing it to lose the component instance and not render anymore.

We get back to that but this case is fixed by adding a listener in `endNavigate` to trigger the component rendering again.

In theory, we get almost the same behavior as adding the tool at the bottom of the body because we avoid the problem with the SPA, but rendering at the bottom causes that we are add the tool script every time in every SPA navigation, and assembles and disassembles the component, that it causes a lot of problems to lose the violations data and the scheduler queue, to take better advantage of the singleton of the scheduler we only load the file once but the component will be remounting every time navigation happens but the scheduler will always be intact, so we don't lose the queue and we managed to have more consistency of the analysis and avoid that at times having more than one scheduler running or operating.