LPS-148445 Rewrite ES modules from ComboServlet
This is a workaround to be able to serve AMD to ESM bridges both directly from
their canonical URL and from combo servlet.

Because we are using relative URLs to refer to ES modules (to avoid issues with
CDN, contexts, etc.) we need to transform them when their contents are served
from /o/combo so that they don't fail.

We also need to transform the variable name used to re-export ES modules through
AMD from esModule to a unique identifier. We use a digest for that to make sure
they don't collide. As these contents are cached, no performance penalty should
be expected.