LPS-183047 Adapt MainServlet to scoping
This call makes inferring the companyId impossible because the thread
locals are not initialized yet, so we pass it explicitly.

AFAIK, at this point there are no company scoped portlets installed because
OSGi registrations have not yet started so I think it is safe to assume
that only the SYSTEM company is available.