LPS-189456 This fixes the double ES module request problem
See https://github.com/liferay-frontend/liferay-portal/pull/3529#issuecomment-1656838090

This makes `frontend-js-components-web/__liferay__/index.js` be requested always
from the same URL (namely that one, without any parameter). This way it is not
treated as different ES modules by the browser.

I've checked that changing the language makes the browser retrieve newer versions
of the ES module while, at the same time, if nothing changes the module is
retrieved from the cache (due to a "340 Not Modified" response being sent from
the server).

I think that's the closest we can get to "forever caching" given the modular
structure of the portal.

I hope this doesn't introduce any regression. It shouldn't, but I've lost all
confidence in changes when it comes to caching modifications... :-(