LPD-25668 Add install task to node-scripts
This is a preliminary task that will be run whenever node-scripts is
installed in the node_modules folder by yarn or npm.

It can be used to implement singleton tasks that benefit from being run as
a setup step of the build instead of being run on-the-fly while executing
any other task.