LPD-44141 Add support for new properties in FDSSortItem and its builder
# breaking

## What modules/apps/frontend-data-set/frontend-data-set-api/src/main/java/com/liferay/frontend/data/set/model/FDSSortItemBuilder.java

This commit adds 2 new methods to FDSSortItem, that allow to set the active flag and the item label. Consequently, FDSSortItemBuilder class added new interfaces and modified some interfaces to accommodate build-related methods handling the new properties.

Particularly, 2 interfaces have been changed:
  public interface AfterDirectionStep now also extends LabelStep interface
  public interface AfterKeyStep now also extends LabelStep interface

## Why

This is needed to accommodate new Data Set definition requirements. It's an internal API not meant to be used outside of DXP.

----