LPD-44142 Provide FDS item action list serializers for custom and system data sets, with unit tests
# breaking

## What modules/apps/frontend-data-set/frontend-data-set-api/src/main/java/com/liferay/frontend/data/set/action/FDSItemActionList.java

deleted method
    * public List<FDSActionDropdownItem> getFDSActionDropdownItems(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse) throws PortalException

added method
    * public List<FDSActionDropdownItem> getFDSActionDropdownItems(
              HttpServletRequest httpServletRequest);

## Why

httpServletResponse is not necessary to compute the item action list. Also, we want implementations to handle their exceptions and return an empty list if something goes wrong, so that the serialization does not stop. This is a recently added interface, not yet implemented. Customers are not affected at the time of this writing

----